+++ cmd/list.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 
@@ -23,18 +22,8 @@ func (a *app) newListCommand() *cobra.Command {
 			if err := cfg.Validate(); err != nil {
 				return i18n.NewError(i18n.ConfigInvalid, a.translator.Error(err))
 			}
-			color := !a.noColor && os.Getenv("NO_COLOR") == "" && isTerminal(a.out)
-			_, err = fmt.Fprint(a.out, presentation.RenderProjects(cfg, color, a.translator))
+			_, err = fmt.Fprint(a.out, presentation.RenderProjects(cfg, a.colorEnabled(), a.translator))
 			return err
 		},
 	}
-}
-
-func isTerminal(w any) bool {
-	file, ok := w.(*os.File)
-	if !ok {
-		return false
-	}
-	info, err := file.Stat()
-	return err == nil && info.Mode()&os.ModeCharDevice != 0
 }